[FIX] No se subían las direcciones

+++ prestadoo/models/res_partner.py
@@ -26,8 +26,7 @@ class Partner(BaseExtClass):
         if self.commercial_partner_id.is_company \
                 and self.commercial_partner_id.customer \
                 and self.commercial_partner_id.email \
-                and self.commercial_partner_id.web_password \
-                and self.group_companies_ids:
+                and self.commercial_partner_id.web_password:
 
             if self.commercial_partner_id.id == self.id:
                 self.fields_to_watch = ('id', 'name', 'vat', 'email', 'web_password', 'property_product_pricelist',
@@ -70,23 +69,26 @@ class Partner(BaseExtClass):
         # El filtro de notificaciones viene hecho ya en "is_notifiable", de forma que aquí llegarán IC's o
         # direcciones de entrega
         if self.commercial_partner_id.id == self.id: # Se trata de un IC
-            # Recorremos la lista de compañías a las que puede tener acceso el IC para averiguar sus tarifas
-            pricelist_list = []
-            valid = 'Y'
-            for company in self.group_companies_ids:
-                if self.env.user.company_id.id == company.id:
-                    pricelist = self.property_product_pricelist
-                else:
-                    pricelist_obj = self.env['product.pricelist']
-                    pricelist_id = pricelist_obj.sudo()._get_partner_pricelist(
-                        self.id, company.id)
-                    pricelist = pricelist_obj.browse(pricelist_id)
-                pricelist_list.append(str(company.id) + ',' + pricelist.legacy_code or str(pricelist.id + 1000))
-                if not unlink and self.active:
-                    if pricelist.name == 'SIN PRECIO':
-                        valid = 'P'
-                else:
-                    valid = 'N'
+            if self.group_companies_ids: # Si está asignado a alguna empresa del grupo
+                # Recorremos la lista de compañías a las que puede tener acceso el IC para averiguar sus tarifas
+                pricelist_list = []
+                valid = 'Y'
+                for company in self.group_companies_ids:
+                    if self.env.user.company_id.id == company.id:
+                        pricelist = self.property_product_pricelist
+                    else:
+                        pricelist_obj = self.env['product.pricelist']
+                        pricelist_id = pricelist_obj.sudo()._get_partner_pricelist(
+                            self.id, company.id)
+                        pricelist = pricelist_obj.browse(pricelist_id)
+                    pricelist_list.append(str(company.id) + ',' + pricelist.legacy_code or str(pricelist.id + 1000))
+                    if not unlink and self.active:
+                        if pricelist.name == 'SIN PRECIO':
+                            valid = 'P'
+                    else:
+                        valid = 'N'
+            else:
+                valid = 'N'
 
             self.xml = pocustomer.format(
                     self.ref or self.id,                   # CardCode
@@ -101,6 +103,11 @@ class Partner(BaseExtClass):
             self.obj_type = '2'
 
         else: # Se trata de una dirección
+            if not unlink and self.active:
+                valid = 'Y'
+            else:
+                valid = 'N'
+
             self.xml = pocustomeraddress.format(
                     self.parent_id.ref or self.parent_id.id,    # CardCode
                     self.legacy_code or '#' + str(self.id),     # LineNum
